Move helper functions to utils; log address when validation skipped in tax apps
